+++ Gruntfile.js
@@ -53,16 +53,62 @@ var BS_BRANCH = process.env.TRAVIS_BRANCH || "local";
 var BS_TAG = process.env.BROWSER_STACK_USERNAME + "@" + BS_BRANCH;
 
 // ----------------------------------------------------------------------------
-// Configuration
+// Globals
 // ----------------------------------------------------------------------------
+// Sauce labs environments.
+var SAUCE_ENVS = {
+  // Already tested in Travis.
+  // sl_firefox: {
+  //   base: "SauceLabs",
+  //   browserName: "firefox"
+  // },
+  sl_chrome: {
+    base: "SauceLabs",
+    browserName: "chrome"
+  },
+  sl_safari: {
+    base: "SauceLabs",
+    browserName: "safari",
+    platform: "OS X 10.9"
+  },
+  sl_ie_9: {
+    base: "SauceLabs",
+    browserName: "internet explorer",
+    platform: "Windows 7",
+    version: "9"
+  },
+  sl_ie_10: {
+    base: "SauceLabs",
+    browserName: "internet explorer",
+    platform: "Windows 7",
+    version: "10"
+  },
+  sl_ie_11: {
+    base: "SauceLabs",
+    browserName: "internet explorer",
+    platform: "Windows 7",
+    version: "11"
+  }
+};
+
+// SauceLabs tag.
+var SAUCE_BRANCH = process.env.TRAVIS_BRANCH || "local";
+var SAUCE_TAG = process.env.SAUCE_USERNAME + "@" + SAUCE_BRANCH;
+
 module.exports = function (grunt) {
 
+  // --------------------------------------------------------------------------
+  // Helpers
+  // --------------------------------------------------------------------------
   // Strip comments from JsHint JSON files.
   var _jshint = function (name) {
     if (!grunt.file.exists(name)) { return "{}"; }
     return JSON.parse(grunt.file.read(name).replace(/\/\/.*\n/g, ""));
   };
 
+  // --------------------------------------------------------------------------
+  // Configuration
+  // --------------------------------------------------------------------------
   grunt.initConfig({
     pkg: grunt.file.readJSON("package.json"),
 
@@ -308,6 +354,21 @@ module.exports = function (grunt) {
         captureTimeout: 0, // Pass through to BS.
         customLaunchers: BS_ENVS,
         browsers: Object.keys(BS_ENVS)
+      },
+      sauce: {
+        singleRun: true,
+        reporters: ["mocha", "saucelabs"],
+        sauceLabs: {
+          testName: "Backbone Testing - Frontend Unit Tests",
+          tags: [SAUCE_TAG],
+          public: "public"
+        },
+        // Timeouts: Allow "n" minutes before saying "good enough". See also:
+        // https://github.com/angular/angular.js/blob/master/
+        //         karma-shared.conf.js
+        captureTimeout: 0, // Pass through to SL.
+        customLaunchers: SAUCE_ENVS,
+        browsers: Object.keys(SAUCE_ENVS)
       }
     },